Make scan throw TransactionClosedError
Also move the ScanResult/ScanIterator to their own file.
