+++ src/fixtures/put, get, has, del inside tx.json
@@ -89,7 +89,7 @@
       "transactionId": 1
     },
     "result": {
-      "ref": "46k84kgj81735i7mj11pqo5o70rp1ltj"
+      "ref": "2cv32fvipvi0kn36lkq2o4ru40263eb8"
     }
   },
   {
@@ -99,14 +99,6 @@
     "result": {
       "transactionId": 2
     }
-  },
-  {
-    "dbName": "test3",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 2
-    },
-    "result": {}
   },
   {
     "dbName": "test3",
@@ -121,6 +113,14 @@
     "result": {
       "transactionId": 3
     }
+  },
+  {
+    "dbName": "test3",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 2
+    },
+    "result": {}
   },
   {
     "dbName": "test3",
@@ -184,7 +184,7 @@
       "transactionId": 3
     },
     "result": {
-      "ref": "jv3u8njdgm7ni43kjvqg7qv363lio9e3"
+      "ref": "cac8csk9rcfec1o3mlc0t7cfm3ivomj0"
     }
   },
   {
@@ -279,7 +279,7 @@
       "transactionId": 5
     },
     "result": {
-      "ref": "ikp56noejc2lnfcjn09a7eeh48u7kvr1"
+      "ref": "8ldvv83o1nrshp23mg4754nem4ec49gn"
     }
   },
   {
@@ -289,6 +289,14 @@
     "result": {
       "transactionId": 6
     }
+  },
+  {
+    "dbName": "test3",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 6
+    },
+    "result": {}
   },
   {
     "dbName": "test3",
@@ -303,14 +311,6 @@
     "result": {
       "transactionId": 7
     }
-  },
-  {
-    "dbName": "test3",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 6
-    },
-    "result": {}
   },
   {
     "dbName": "test3",
@@ -374,7 +374,7 @@
       "transactionId": 7
     },
     "result": {
-      "ref": "2l5ls00kib061urolkomj2sjstirot6t"
+      "ref": "vrf6l81rjjo80ejehtlkkr98iq6d2gbs"
     }
   },
   {
@@ -469,7 +469,7 @@
       "transactionId": 9
     },
     "result": {
-      "ref": "vbackrcava8e28629qgp4h8eu1bkk725"
+      "ref": "dtv73vb7hf8m1oimhfvqhqsasbsn82d4"
     }
   },
   {
@@ -479,6 +479,14 @@
     "result": {
       "transactionId": 10
     }
+  },
+  {
+    "dbName": "test3",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 10
+    },
+    "result": {}
   },
   {
     "dbName": "test3",
@@ -493,14 +501,6 @@
     "result": {
       "transactionId": 11
     }
-  },
-  {
-    "dbName": "test3",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 10
-    },
-    "result": {}
   },
   {
     "dbName": "test3",
@@ -564,7 +564,7 @@
       "transactionId": 11
     },
     "result": {
-      "ref": "msbs7dhtq1mdmjkspdmlgl1c27b25t3q"
+      "ref": "rpr4rlbv7ppes9165tfs5ogdaqi6ap5i"
     }
   },
   {
@@ -574,6 +574,14 @@
     "result": {
       "transactionId": 12
     }
+  },
+  {
+    "dbName": "test3",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 12
+    },
+    "result": {}
   },
   {
     "dbName": "test3",
@@ -588,14 +596,6 @@
     "result": {
       "transactionId": 13
     }
-  },
-  {
-    "dbName": "test3",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 12
-    },
-    "result": {}
   },
   {
     "dbName": "test3",
@@ -659,7 +659,7 @@
       "transactionId": 13
     },
     "result": {
-      "ref": "2gqq97gmnaloul25r2ahqt8kbo7cgm5o"
+      "ref": "stuvecgd024s7ktrcn405327t5shf16e"
     }
   },
   {
@@ -760,7 +760,7 @@
       "transactionId": 15
     },
     "result": {
-      "ref": "phhiqcck74plf5imiofqv66bsl35114l"
+      "ref": "57shmf9caa15v6slgjs07lcfdiavuesk"
     }
   },
   {
@@ -770,6 +770,14 @@
     "result": {
       "transactionId": 16
     }
+  },
+  {
+    "dbName": "test3",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 16
+    },
+    "result": {}
   },
   {
     "dbName": "test3",
@@ -787,14 +795,6 @@
     "result": {
       "transactionId": 17
     }
-  },
-  {
-    "dbName": "test3",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 16
-    },
-    "result": {}
   },
   {
     "dbName": "test3",
@@ -864,7 +864,7 @@
       "transactionId": 17
     },
     "result": {
-      "ref": "bovoshir9v6jb1132gdqfo9loqpuv6vl"
+      "ref": "bqmk1odl4frdl3jd3l8e7jmr30ek49kq"
     }
   },
   {
