+++ src/fixtures/scan.json
@@ -131,7 +131,7 @@
       "transactionId": 1
     },
     "result": {
-      "ref": "lisqeb2kmo0hme1d43t0vogrml253ho7"
+      "ref": "r62js7d1k87vd5ds6uaai6h537kgd3ug"
     }
   },
   {
@@ -335,6 +335,14 @@
         "value": 8
       }
     ]
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 4
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -346,14 +354,6 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 4
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "scan",
     "args": {
       "transactionId": 5,
@@ -432,6 +432,14 @@
         "value": 8
       }
     ]
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 5
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -440,14 +448,6 @@
     "result": {
       "transactionId": 6
     }
-  },
-  {
-    "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 5
-    },
-    "result": {}
   },
   {
     "dbName": "test4",
@@ -657,14 +657,6 @@
         "value": 4
       }
     ]
-  },
-  {
-    "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 8
-    },
-    "result": {}
   },
   {
     "dbName": "test4",
@@ -673,6 +665,14 @@
     "result": {
       "transactionId": 9
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 8
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -827,19 +827,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 11
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 12
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 11
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -905,19 +905,19 @@
   },
   {
     "dbName": "test4",
-    "method": "openTransaction",
-    "args": {},
-    "result": {
-      "transactionId": 14
-    }
-  },
-  {
-    "dbName": "test4",
     "method": "closeTransaction",
     "args": {
       "transactionId": 13
     },
     "result": {}
+  },
+  {
+    "dbName": "test4",
+    "method": "openTransaction",
+    "args": {},
+    "result": {
+      "transactionId": 14
+    }
   },
   {
     "dbName": "test4",
@@ -944,19 +944,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 14
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 15
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 14
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -1068,19 +1068,19 @@
   },
   {
     "dbName": "test4",
-    "method": "openTransaction",
-    "args": {},
-    "result": {
-      "transactionId": 19
-    }
-  },
-  {
-    "dbName": "test4",
     "method": "closeTransaction",
     "args": {
       "transactionId": 18
     },
     "result": {}
+  },
+  {
+    "dbName": "test4",
+    "method": "openTransaction",
+    "args": {},
+    "result": {
+      "transactionId": 19
+    }
   },
   {
     "dbName": "test4",
@@ -1113,19 +1113,19 @@
   },
   {
     "dbName": "test4",
-    "method": "openTransaction",
-    "args": {},
-    "result": {
-      "transactionId": 20
-    }
-  },
-  {
-    "dbName": "test4",
     "method": "closeTransaction",
     "args": {
       "transactionId": 19
     },
     "result": {}
+  },
+  {
+    "dbName": "test4",
+    "method": "openTransaction",
+    "args": {},
+    "result": {
+      "transactionId": 20
+    }
   },
   {
     "dbName": "test4",
@@ -1158,19 +1158,19 @@
   },
   {
     "dbName": "test4",
-    "method": "openTransaction",
-    "args": {},
-    "result": {
-      "transactionId": 21
-    }
-  },
-  {
-    "dbName": "test4",
     "method": "closeTransaction",
     "args": {
       "transactionId": 20
     },
     "result": {}
+  },
+  {
+    "dbName": "test4",
+    "method": "openTransaction",
+    "args": {},
+    "result": {
+      "transactionId": 21
+    }
   },
   {
     "dbName": "test4",
@@ -1248,19 +1248,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 22
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 23
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 22
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -1289,19 +1289,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 23
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 24
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 23
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -1412,19 +1412,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 26
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 27
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 26
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
@@ -1538,19 +1538,19 @@
   },
   {
     "dbName": "test4",
-    "method": "closeTransaction",
-    "args": {
-      "transactionId": 29
-    },
-    "result": {}
-  },
-  {
-    "dbName": "test4",
     "method": "openTransaction",
     "args": {},
     "result": {
       "transactionId": 30
     }
+  },
+  {
+    "dbName": "test4",
+    "method": "closeTransaction",
+    "args": {
+      "transactionId": 29
+    },
+    "result": {}
   },
   {
     "dbName": "test4",
