+++ src/fixtures/subscribe with error.json
@@ -70,7 +70,7 @@
       "transactionId": 2
     },
     "result": {
-      "ref": "udnd1vcj7b4d5mqavrl59ob7gss3pm8b"
+      "ref": "fuvlabho49l1kklklp7ifok78j68p58k"
     }
   },
   {
