+++ src/fixtures/subscribe.json
@@ -69,7 +69,7 @@
       "transactionId": 2
     },
     "result": {
-      "ref": "dca21o1kqmcu7v566og35m89t6am61dr"
+      "ref": "1urd54h8eang8qlg04a2252jdensto5m"
     }
   },
   {
@@ -133,7 +133,7 @@
       "transactionId": 4
     },
     "result": {
-      "ref": "8u801jor8drm5hc1piliv130k027fbg4"
+      "ref": "hu07btpggcmkc4ujo1bmlvm9o6li3bh0"
     }
   },
   {
@@ -197,7 +197,7 @@
       "transactionId": 6
     },
     "result": {
-      "ref": "9953q0p3d0cso6ubslqtidgcpl9mbf8s"
+      "ref": "rrv3aqst7so8e9a6kalic5s9etm95dv9"
     }
   },
   {
@@ -265,7 +265,7 @@
       "transactionId": 8
     },
     "result": {
-      "ref": "7hu48ljvcjj3nrif48d6pbe4fmbm1hqd"
+      "ref": "r84c27uhge22td3eos9547jrjpch7ge5"
     }
   },
   {
@@ -333,7 +333,7 @@
       "transactionId": 10
     },
     "result": {
-      "ref": "70f4cu26f8lpnaq0pk4uafcn4lg35o75"
+      "ref": "7vnuku1ds84fd43c7bk4mp5848ealr55"
     }
   },
   {
