+++ src/mod.ts
@@ -1,12 +1,11 @@
 export {default} from './replicache.js';
+export {TransactionClosedError} from './transaction-closed-error.js';
+export {REPMHTTPInvoker} from './repm-invoker.js';
+
 export type {Mutator} from './replicache.js';
 export type {ScanId} from './scan-id.js';
 export type {ScanBound} from './scan-bound.js';
 export type {DatabaseInfo} from './database-info.js';
-export {REPMHTTPInvoker} from './repm-invoker.js';
 export type {REPMInvoke} from './repm-invoker.js';
-export type {
-  ReadTransaction,
-  WriteTransaction,
-  ScanResult,
-} from './transactions.js';
+export type {ReadTransaction, WriteTransaction} from './transactions.js';
+export type {ScanResult} from './scan-iterator.js';
