+++ src/transaction-closed-error.ts
@@ -1,0 +1,5 @@
+export class TransactionClosedError extends Error {
+  constructor() {
+    super('Transaction is closed');
+  }
+}
